backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
PyJWT==2.10.1
python-multipart==0.0.17
alembic==1.14.0
